+++ src/bin/rdcore/kargs.rs
@@ -57,7 +57,6 @@ pub fn kargs(config: KargsConfig) -> Result<()> {
                 mount.mountpoint(),
                 None,
                 None,
-                None,
                 s390x::ZiplSecexMode::Auto,
             )?;
         }
@@ -98,7 +97,6 @@ pub fn zipl(config: ZiplConfig) -> Result<()> {
     s390x::zipl(
         boot.mountpoint(),
         config.hostkey,
-        config.rootfs,
         config.append_karg.as_ref().map(|v| v.join(" ")),
         config.secex_mode,
     )