+++ docs/release-notes.md
@@ -27,6 +27,7 @@ Internal changes:
 - zipl: Deprecate `--kargs` in favor of new `--append-karg`, which supports multiple instances
 - Only open BLS configs in write mode when modifying
 - zipl: Support Secure Execution systems without LUKS for coreos-assembler
+- zipl: Remove obsolete `--rootfs`
 
 Packaging changes: